+++ src/core/scene/components/particle/ParticleManagerEditorComponent.cpp
@@ -46,23 +46,23 @@ void ParticleManagerEditorComponent::drawParticleManager()
             particleManager->setCullingWorldSize({ val[0], val[1], val[2] });
         });
 
-        std::array numLayer = { particleManager->getCullingLayerNumber() };
-        m_particleManagerGroup->createWidget<Drag<int>>("Number Layers", ImGuiDataType_S32, numLayer, 1, 1, 8)->getOnDataChangedEvent().addListener([this](auto val) {
+        std::array numLayer = { (float)particleManager->getCullingLayerNumber() };
+        m_particleManagerGroup->createWidget<Drag<float>>("Number Layers", ImGuiDataType_S32, numLayer, 1, 1, 8)->getOnDataChangedEvent().addListener([this](auto val) {
             auto particleManager = std::dynamic_pointer_cast<ParticleManager>(getComponent<CompoundComponent>()->getComponents()[0]);
-            particleManager->setCullingLayerNumber(val[0]);
+            particleManager->setCullingLayerNumber((int)val[0]);
         });
     }
 
-    std::array maxParticles = { particleManager->getMaxParticleNumber() };
-    m_particleManagerGroup->createWidget<Drag<int>>("Max Parcicles", ImGuiDataType_S32, maxParticles, 1, 1)->getOnDataChangedEvent().addListener([this](auto val) {
+    std::array maxParticles = { (float)particleManager->getMaxParticleNumber() };
+    m_particleManagerGroup->createWidget<Drag<float>>("Max Parcicles", ImGuiDataType_S32, maxParticles, 1, 1)->getOnDataChangedEvent().addListener([this](auto val) {
         auto particleManager = std::dynamic_pointer_cast<ParticleManager>(getComponent<CompoundComponent>()->getComponents()[0]);
-        particleManager->setMaxParticleNumber(val[0]);
+        particleManager->setMaxParticleNumber((int)val[0]);
     });
 
-    std::array threadNum = { particleManager->getNumberOfThreads() };
-    m_particleManagerGroup->createWidget<Drag<int>>("Number Threads", ImGuiDataType_S32, threadNum, 1, 1, 4)->getOnDataChangedEvent().addListener([this](auto val) {
+    std::array threadNum = { (float)particleManager->getNumberOfThreads() };
+    m_particleManagerGroup->createWidget<Drag<float>>("Number Threads", ImGuiDataType_S32, threadNum, 1, 1, 4)->getOnDataChangedEvent().addListener([this](auto val) {
         auto particleManager = std::dynamic_pointer_cast<ParticleManager>(getComponent<CompoundComponent>()->getComponents()[0]);
-        particleManager->setNumberOfThreads(val[0]);
+        particleManager->setNumberOfThreads((int)val[0]);
     });
 }
 NS_IGE_END
